feat(combobox): rework combobox ui5 flaw seamless auto detection for control class§
