+++ examples/ui5-js-app/webapp/controller/Main.controller.js
@@ -28,9 +28,7 @@ sap.ui.define(
             initCombobox: function () {
                 // set explored app's demo model on this sample
                 var oModel = new JSONModel()
-                oModel.loadData(
-                    "https://raw.githubusercontent.com/SAP/openui5/master/src/sap.ui.documentation/test/sap/ui/documentation/sdk/countriesExtendedCollection.json"
-                )
+                oModel.loadData("model/countries.json")
                 this.getView().setModel(oModel, "Countries")
             },
