+++ examples/ui5-js-app/webapp/test/e2e/basic.test.js
@@ -1,15 +1,5 @@
 const Main = require("./pageObjects/Main")
 
-const oComboboxSelector = {
-    // forceSelect: true,
-    controlType: "sap.m.ComboBox", // this invokes some internal special case hanlding to fir the ui5 Bug https://github.com/SAP/openui5/issues/3477
-    selector: {
-        interaction: "root",
-        id: "combobox",
-        viewName: "test.Sample.view.Main"
-    }
-}
-
 describe("ui5 basic", () => {
     before(async () => {
         await Main.open()
@@ -59,24 +49,6 @@ describe("ui5 basic", () => {
         // check if result contains the expected validation error
         expect(invalidControl1).toContain("ERROR")
         expect(invalidControl2).toContain("ERROR")
-    })
-
-    // #121
-    it("get combox items aggregation as WebdriverIO representations", async () => {
-        const combobox = await browser.asControl(oComboboxSelector)
-
-        const items = await combobox.getItems(true)
-        expect(items.length).toEqual(70)
-    })
-
-    // #121
-    it("get combox items aggregation as WebdriverIO representations", async () => {
-        const combobox = await browser.asControl(oComboboxSelector)
-        // another issue with the combobox. If the Box was not opend prevoiusly the items are not rendered -> unretrieveable with ui5
-        await combobox.open()
-
-        const items = await combobox.getItems(4)
-        expect(await items.getTitle()).toEqual("Bahrain")
     })
 
     /* it("check for Searchfield Properties", async () => {
